+++ libs/core/src/lib/layout-panel/layout-panel-footer/layout-panel-footer.component.ts
@@ -1,24 +1,52 @@
-import { ChangeDetectionStrategy, Component, HostBinding, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, ElementRef, Input, OnInit, ViewEncapsulation } from '@angular/core';
+import { applyCssClass } from '@fundamental-ngx/cdk/utils';
+import { CssClassBuilder } from '@fundamental-ngx/cdk/utils';
+import { Nullable } from '@fundamental-ngx/cdk/utils';
 
-/**
- * Layout Panel footer can be utilized for pagination, secondary actions, add more data, etc.
- *
- * ```html
- * <fd-layout-panel>
- *     <fd-layout-panel-footer>
- *         Some text can go here!
- *     </fd-layout-panel-footer>
- * </fd-layout-panel>
- * ```
- */
+export type FooterPosition = 'start' | 'end';
+
 @Component({
     selector: 'fd-layout-panel-footer',
     templateUrl: './layout-panel-footer.component.html',
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class LayoutPanelFooterComponent {
+export class LayoutPanelFooterComponent implements OnInit, CssClassBuilder {
+    /** Apply user custom styles */
+    @Input()
+    class: string;
+
+    /**
+     * Footer variations, can be start (left aligned), end (right aligned) or null for default.
+     * The default value will render the content of the footer in the middle.
+     */
+    @Input() position: Nullable<FooterPosition>;
+
     /** @hidden */
-    @HostBinding('class.fd-layout-panel__footer')
-    fdLayoutPanelFooterClass = true;
+    constructor(private readonly _elementRef: ElementRef<HTMLElement>) {}
+
+    /** @hidden */
+    ngOnInit(): void {
+        this.buildComponentCssClass();
+    }
+
+    /** @hidden */
+    ngOnChanges(): void {
+        this.buildComponentCssClass();
+    }
+
+    /** @hidden */
+    @applyCssClass
+    buildComponentCssClass(): string[] {
+        return [
+            'fd-layout-panel__footer',
+            this.position ? `fd-layout-panel__footer--${this.position}` : '',
+            this.class
+        ];
+    }
+
+    /** @hidden */
+    elementRef(): ElementRef<HTMLElement> {
+        return this._elementRef;
+    }
 }