+++ libs/docs/platform/standard-list-item/examples/platform-standard-list-item-border-less-example.component.ts
@@ -1,17 +1,24 @@
 import { Component } from '@angular/core';
+import { ListAvatarConfig } from '@fundamental-ngx/platform/list';
+
+export interface ListItem {
+    title: string;
+    avatar: Partial<ListAvatarConfig>;
+    description: string;
+    secondary: string;
+}
 
 @Component({
     selector: 'fdp-platform-standard-list-item-border-less-example',
     templateUrl: './platform-standard-list-item-border-less-example.component.html'
 })
 export class PlatformStandardListItemBorderLessExampleComponent {
-    items: any[] = [
+    items: ListItem[] = [
         {
             title:
                 'Item1 A laptop computer, sometimes called a notebook computer by manufacturers,' +
                 'is a battery- or AC-powered personal computer generally smaller than a briefcase that can ' +
                 'easily be transported and conveniently',
-            avatarTitle: 'nature',
             description:
                 'First text item in Byline (Standard text item)' +
                 'A laptop computer, sometimes called a notebook computer by manufacturers,' +
@@ -22,26 +29,36 @@ export class PlatformStandardListItemBorderLessExampleComponent {
                 'A laptop computer, sometimes called a notebook computer by manufacturers,' +
                 'is a battery- or AC-powered personal computer generally smaller than a briefcase that can ' +
                 'easily be transported and conveniently',
-            avatarSrc: 'https://placeimg.com/400/400/nature'
+            avatar: {
+                image: 'https://placeimg.com/400/400/nature',
+                ariaLabel: 'nature'
+            }
         },
         {
             title: 'Item2',
-            avatarSrc: 'https://placeimg.com/400/400/nature',
-            avatarTitle: 'nature',
             description: 'First text item in Byline (Standard text item)',
-            secondary: 'Second text item in Byline (Can be semantic (Status) or not)'
+            secondary: 'Second text item in Byline (Can be semantic (Status) or not)',
+            avatar: {
+                circle: true,
+                glyph: 'account',
+                colorAccent: 10
+            }
         },
         {
             title: 'Item3',
-            avatarSrc: 'https://placeimg.com/400/400/nature',
-            avatarTitle: 'nature',
             description: 'First text item in Byline (Standard text item)',
-            secondary: 'Second text item in Byline (Can be semantic (Status) or not)'
+            secondary: 'Second text item in Byline (Can be semantic (Status) or not)',
+            avatar: {
+                image: 'https://placeimg.com/400/400/nature',
+                ariaLabel: 'nature'
+            }
         },
         {
             title: 'Item4',
-            avatarSrc: 'https://placeimg.com/400/400/nature',
-            avatarTitle: 'nature',
+            avatar: {
+                image: 'https://placeimg.com/400/400/nature',
+                ariaLabel: 'nature'
+            },
             description: 'First text item in Byline (Standard text item)',
             secondary: 'Second text item in Byline (Can be semantic (Status) or not)'
         }